Better sorting and rename of file/folders
