+++ client/src/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import '../src/components/styles/App.css';
+import './styles/app.css';
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Home from "./components/pages/Home.js";
-import Setup from "./components/pages/Setup.js";
-import User from "./components/pages/User.js";
-import Choices from "./components/pages/Choices.js";
+import Home from "./pages/Home.js";
+import Signup from "./pages/SignUp.js";
+import User from "./pages/User.js";
+import Choices from "./pages/Choices.js";
 
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
     <Router>
       <div>
         <Route exact path="/" component={Home} />
-        <Route exact path="/setup" component={Setup} />
+        <Route exact path="/signup" component={Signup} />
         <Route exact path="/user" component={User} />
         <Route exact path="/choices" component={Choices} />
       </div>
