+++ client/src/components/Footer/Footer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import "./styles/Footer.css";
+import "./style.css";
 
 export default function Footer() {
     return (
