+++ client/src/components/Swipe/Swipe.js
