+++ client/src/pages/Choices.js
